small style updates on pm, result and category pages

+++ app/controllers/PMController.php
@@ -255,8 +255,10 @@ class PMController extends BaseController {
 			}
 		}
 		$pm->title = Input::get('title');
-		$pm->categories()->detach();
-		$pm->categories()->attach([Input::get('category') => ['added_by' => Auth::user()->id]]);
+		if(Input::get('category') != 0) {
+			$pm->categories()->detach();
+			$pm->categories()->attach([Input::get('category') => ['added_by' => Auth::user()->id]]);
+		}
 		$pm->content = Input::get('content');
 		$pm->save();
